refactor: relocate quiz types
