+++ components/TakeAnswer/MultiSelectAnswer.tsx
@@ -5,7 +5,7 @@ import { Label } from "@/components/ui/label";
 import { Check, CheckCircle, X, XCircle } from "lucide-react";
 import { useEffect, useState } from 'react';
 
-import { Answer } from '@/component-models/types';
+import { Answer } from '@/types/quiz';
 
 import { get } from '../../services/HttpService';
