+++ components/TakeAnswer/SingleSelectAnswer.tsx
@@ -5,7 +5,7 @@ import { RadioGroup, RadioGroupItem } from "@/components/ui/radio-group";
 import { CheckCircle, X, XCircle } from "lucide-react";
 import { useEffect, useState } from 'react';
 
-import { Answer } from '@/component-models/types';
+import { Answer } from '@/types/quiz';
 
 import { get } from '../../services/HttpService';
