+++ types/quiz.ts
